requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.105.0
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3.2
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.42
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
streamlit==1.30.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
wheel==0.42.0
zipp==3.17.0
